splitting with a measure can return the summary of the split chunks
This allows us to avoid recalculating the summary of the left and right
chunks by providing the original summary. Often times the individual
split implementations on a measure can compute the final summary from
the initial one much faster than `slice.summarize()`.

This brings a ~21% improvement to the `lines_tiny` benchmark, a 37%
improvement to `lines_large` but also a 37% regression on `iter/chars`.

TODO: avoid the regression in the chars iterator while keeping the
performance gains obtained for the lines iterator.