fix `Chops::next`
fixes a few stack-related logic bugs in the implementation of
`Chops::next` that was causing some tests on the `Lines` iterator to
fail.

It also refactors `Chops` to use a `VecDeque` instead of a `Vec` for the
stack. This simplifies the `Chops::next` implementation and also makes
it so that the future `Chops::next_back` will have similar performance
characteristics.

However it also introduces a regression in the `iter_lines_forward`
benchmarks. In particular `tiny` is 16% slower and `large` is 43%
slower :(.