refactor output of `Metric::split_left`
Now the optional right part is combined inside a tuple inside the
`Option`, and the left summary is no longer optional.

This allows to avoid an `unwrap_or` at the call sites, and since
`L::Slice`s are `Summarize` they can always be summarized inside the
`Metric` implementations if needed.

This brings a 17% improvement in the `iter_lines_forward/lines_tiny`
benchmark and a 31% improvement in `iter_lines_forward/lines_large`,
pretty much getting back to the original speed before the last commit,
except now `Chops::next` actually works and it already uses a
`VecDeque` :).